pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-app"
version = "0.1.0"
description = "Minimal agentic math tool router using an LLM and CLI interface."
requires-python = ">=3.11"
dependencies = [
    "openai>=1.12.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
agentic-math = "agentic_app.app:main"


